% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_packages.R
\name{recommend_packages}
\alias{recommend_packages}
\title{Recommend Packages Based on File Extensions}
\usage{
recommend_packages(files_df)
}
\arguments{
\item{files_df}{A data frame containing file names and extensions, typically the output from \code{scan_folder}.}
}
\value{
A data frame with an additional column:
\describe{
\item{Recommendation}{Suggested R packages or methods for each file type.}
}
}
\description{
This function provides recommendations for R packages or methods to open or analyze files based on their extensions.
}
\examples{
\dontrun{
recommendations_df <- recommend_packages(files_df)
}

}
